==> master_thesis_hypertension-complications-developer/Cohort_Extraction/2.Complications_Cohorts_CR.ipynb <==
import fiber
from fiber.cohort import Cohort
from fiber.condition import Patient, MRNs
from fiber.condition import Diagnosis
from fiber.condition import Measurement, Encounter, Drug
from fiber.storage import yaml as fiberyaml
import pandas as pd
import pyarrow.parquet as pq
import numpy as np
import os
from functools import reduce 
from fiber.utils import Timer
import pickle

#Cases with disease development cohort extraction

#Load dataframe HT cohort
Cases_HT_beforeOnset2013= pd.read_pickle("./Cases_HT_beforeOnset2013.pkl")
Cases_HT_beforeOnset2013 = Cases_HT_beforeOnset2013.rename(columns={"onset_year": "HT_onset_year"})
Cases_HT_beforeOnset2013 = Cases_HT_beforeOnset2013.rename(columns={"age_in_days": "age_in_days_HT_onset"})

def df_to_cohort(df):
    mrns = list(df.index.values)
    condition = MRNs(mrns)
    return Cohort(condition)

cohort = df_to_cohort(Cases_HT_beforeOnset2013)

#stroke/Cerebrovascular diseases
condition = (Diagnosis('I60.%', 'ICD-10')|
             Diagnosis('I61.%', 'ICD-10')|
             Diagnosis('I62.%', 'ICD-10')|
             Diagnosis('I63.%', 'ICD-10')|
             Diagnosis('I64.%', 'ICD-10')|
             Diagnosis('I65.%', 'ICD-10')|
             Diagnosis('I66.%', 'ICD-10')|
             Diagnosis('I67.%', 'ICD-10')|
             Diagnosis('I68.%', 'ICD-10')|
             Diagnosis('I69.%', 'ICD-10')|
             Diagnosis('G45.%', 'ICD-10')|
             Diagnosis('430.%', 'ICD-9')|
             Diagnosis('431.%', 'ICD-9')|
             Diagnosis('432.%', 'ICD-9')|
             Diagnosis('433.%', 'ICD-9')|
             Diagnosis('434.%', 'ICD-9')|
             Diagnosis('435.%', 'ICD-9')|
             Diagnosis('436.%', 'ICD-9')|
             Diagnosis('437.%', 'ICD-9')|
             Diagnosis('438.%', 'ICD-9')|
            #heart disease
             Diagnosis('I20.%', 'ICD-10')|
             Diagnosis('I21.%', 'ICD-10')|
             Diagnosis('I22.%', 'ICD-10')|
             Diagnosis('I23.%', 'ICD-10')|
             Diagnosis('I24.%', 'ICD-10')|
             Diagnosis('I25.%', 'ICD-10')|
             Diagnosis('I26.%', 'ICD-10')|
             Diagnosis('I27.%', 'ICD-10')|
             Diagnosis('I28.%', 'ICD-10')|
             Diagnosis('I30.%', 'ICD-10')|
             Diagnosis('I31.%', 'ICD-10')|
             Diagnosis('I32.%', 'ICD-10')|
             Diagnosis('I33.%', 'ICD-10')|
             Diagnosis('I34.%', 'ICD-10')|
             Diagnosis('I35.%', 'ICD-10')|
             Diagnosis('I36.%', 'ICD-10')|
             Diagnosis('I37.%', 'ICD-10')|
             Diagnosis('I38.%', 'ICD-10')|
             Diagnosis('I39.%', 'ICD-10')|
             Diagnosis('I40.%', 'ICD-10')|
             Diagnosis('I41.%', 'ICD-10')|
             Diagnosis('I42.%', 'ICD-10')|
             Diagnosis('I43.%', 'ICD-10')|
             Diagnosis('I44.%', 'ICD-10')|
             Diagnosis('I45.%', 'ICD-10')|
             Diagnosis('I46.%', 'ICD-10')|
             Diagnosis('I47.%', 'ICD-10')|
             Diagnosis('I48.%', 'ICD-10')|
             Diagnosis('I49.%', 'ICD-10')|
             Diagnosis('I50.%', 'ICD-10')|
             Diagnosis('I51.%', 'ICD-10')|
             Diagnosis('I52.%', 'ICD-10')|
             Diagnosis('410.%', 'ICD-9')|
             Diagnosis('411.%', 'ICD-9')|
             Diagnosis('412.%', 'ICD-9')|
             Diagnosis('413.%', 'ICD-9')|
             Diagnosis('414.%', 'ICD-9')|
             Diagnosis('415.%', 'ICD-9')|
             Diagnosis('416.%', 'ICD-9')|
             Diagnosis('417.%', 'ICD-9')|
             Diagnosis('420.%', 'ICD-9')|
             Diagnosis('421.%', 'ICD-9')|
             Diagnosis('422.%', 'ICD-9')|
             Diagnosis('423.%', 'ICD-9')|
             Diagnosis('424.%', 'ICD-9')|
             Diagnosis('425.%', 'ICD-9')|
             Diagnosis('426.%', 'ICD-9')|
             Diagnosis('427.%', 'ICD-9')|
             Diagnosis('428.%', 'ICD-9')|
             Diagnosis('I11.%', 'ICD-10')| #hypertensive heart disease
             Diagnosis('402.%', 'ICD-9')|  #hypertensive heart disease
             Diagnosis('429.%', 'ICD-9')|
             #renal disease
             Diagnosis('N17.%', 'ICD-10')|
             Diagnosis('N18.%', 'ICD-10')|
             Diagnosis('N19.%', 'ICD-10')|
             Diagnosis('I12.%', 'ICD-10')| #hypertensive kidney
             Diagnosis('584.%', 'ICD-9')|
             Diagnosis('585.%', 'ICD-9')|
             Diagnosis('403.%', 'ICD-9')| #hypertensive kidney
             Diagnosis('586.%', 'ICD-9'))
  
Cases_HT_before2013_disease = cohort.get(condition)

Cases_HT_before2013_disease.set_index('medical_record_number', inplace = True)

### create new dataframe with only disease MRN and date_of_birth
Cases_HT_beforeOnset2013_DOB = Cases_HT_beforeOnset2013[['date_of_birth_actual','HT_onset_year', 'age_in_days_HT_onset']]
Cases_HT_before2013_disease_merged_DOB = Cases_HT_before2013_disease.merge(Cases_HT_beforeOnset2013_DOB, left_index=True, right_index=True, how="inner")

#add onset year
def add_onset_year(df_main):
    df_main['age_in_days_delta'] = pd.to_timedelta(df_main['age_in_days'],'d')
    df_main['onset_year'] = (df_main['date_of_birth_actual'] + df_main['age_in_days_delta']).dt.year
    return df_main

Cases_HT_before2013_disease_onset = add_onset_year(Cases_HT_before2013_disease_merged_DOB)
Cases_HT_before2013_disease_onset_unique = Cases_HT_before2013_disease_onset.groupby("medical_record_number").count()


Cases_HT_before2013_disease_onset

Cases_HT_before2013_disease_onset_unique

#save as pkl files
Cases_HT_before2013_disease_onset.to_pickle("./Cases_HT_all3_cohorts_all.pkl")
Cases_HT_before2013_disease_onset_unique.to_pickle("./Cases_HT_all3_cohorts_unique.pkl")

# Control Creation

#Choose file to pick from
Cases_complication = pd.read_pickle("./Cases_HT_all3_cohorts_unique.pkl")

#Create file with only MRNs for the cases
Cases_complication_OnlyMRNs = Cases_complication.drop(['age_in_days', 'context_name', 'context_diagnosis_code','date_of_birth_actual','age_in_days_delta', 'onset_year', 'HT_onset_year','age_in_days_HT_onset'], axis = 1) 

#Merge All Cases with only CR_MRNS
Controls_without_complication = Cases_HT_beforeOnset2013.merge(Cases_complication_OnlyMRNs, on ='medical_record_number', how='outer', indicator=True)

#Create new dataframe where MRNs that appear in both dataframes before are ignored
Controls = Controls_without_complication[Controls_without_complication._merge != 'both']


Controls

#name file correctly and save cohort
Controls.to_pickle("./Controls_No_all3_cohorts_Disease_Ever.pkl")
