==> master_thesis_hypertension-complications-developer/Cohort_Extraction/3.Complications_after_HT_diagnosed.ipynb <==
import pandas as pd
import pyarrow.parquet as pq
import numpy as np
import os
from functools import reduce 
import pickle

#Load Dataframe of 

All_Disease_Cases = pd.read_pickle('/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/All3_diseases/All/Cases_HT_all3_cohorts_all.pkl')
#All_Disease_Cases = pd.read_pickle('/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/Cerebrovascular_Diseases/All/Cases_HT_CR_all.pkl')
#All_Disease_Cases = pd.read_pickle('/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/Heart_Diseases/All/Cases_HT_heart_disease_all.pkl')
#All_Disease_Cases = pd.read_pickle('/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/Renal_Diseases/All/Cases_HT_renal_disease_all.pkl')

###
#All_Disease_Cases = pd.read_pickle('Complications/CR_And_Heart/All/Cases_HT_CR_AND_Heart_all.pkl')
#old Cohorts
    #All_Disease_Cases = pd.read_pickle('Complications/Arterial_Diseases/Cases_HT_arterial_disease_all.pkl')
    #All_Disease_Cases = pd.read_pickle('Complications/Dementia/Cases_HT_dementia_all.pkl')
    #All_Disease_Cases = pd.read_pickle('Complications/Hypertensive_Diseases/Cases_HT_hypertensive_disease_all.pkl')


All_Disease_Cases.head()

#Disease cases before HT_Onset 
Disease_Cases_before_HT_onset = All_Disease_Cases.loc[All_Disease_Cases['age_in_days_HT_onset'] >= All_Disease_Cases['age_in_days']]

#save unique mrns for those candidates that had disease before HT_onset
Disease_Cases_before_HT_onset_unique= Disease_Cases_before_HT_onset.groupby("medical_record_number").count()

### NEWLY diagnosed - First time disease only cases that have been NEWLY diagnosed with disease after HT_Onset
#index those MRNs that have been diagnosed with the disease already before their HT_onset and drop MRNs from entire disease set

indexNames= Disease_Cases_before_HT_onset_unique.index
All_Disease_Cases.drop(indexNames,inplace = True)

#rename cohort
Disease_Cases_after_HT_onset = All_Disease_Cases

#identify unique mrns that have been diagnosed for the first time with that disease and after their HT_Onset
Disease_Cases_after_HT_onset_unique = Disease_Cases_after_HT_onset.groupby("medical_record_number").count()



Disease_Cases_before_HT_onset_unique

Disease_Cases_after_HT_onset

Disease_Cases_after_HT_onset.loc[Disease_Cases_after_HT_onset['age_in_days_HT_onset'] >= Disease_Cases_after_HT_onset['age_in_days']]

Disease_Cases_after_HT_onset_unique

#save as pickle files
#Disease_Cases_after_HT_onset.to_pickle("/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/Renal_Diseases/Cases_renal_disease_after_HT_onset.pkl")
#Disease_Cases_after_HT_onset_unique.to_pickle("/home/kiwitn01/master_thesis_hypertension-complications/Cohort_Extraction/Complications/Renal_Diseases/Cases_renal_disease_after_HT_onset_unique.pkl")
